--- a/go.mod
+++ b/go.mod
@@ -2,6 +2,8 @@ module github.com/influxdata/flux
 
 go 1.12
 
+replace github.com/influxdata/tdigest => /home/flux/code/tdigest
+
 require (
 	cloud.google.com/go v0.52.0
 	cloud.google.com/go/bigtable v1.3.0