--- a/stdlib/universe/quantile.go
+++ b/stdlib/universe/quantile.go
@@ -218,7 +218,10 @@ func createQuantileTransformation(id execute.DatasetID, mode execute.Accumulatio
 		return nil, nil, errors.Newf(codes.Internal, "invalid spec type %T", ps)
 	}
 	agg := NewQuantileAgg(ps.Quantile, ps.Compression)
-	_ = a.Allocator().Account(tdigest.ByteSizeForCompression(agg.Compression))
+	err := a.Allocator().Account(tdigest.ByteSizeForCompression(agg.Compression))
+	if err != nil {
+		return nil, nil, errors.Newf(codes.Internal, "could not allocate memory for tdigest: %s", err)
+	}
 	t, d := execute.NewAggregateTransformationAndDataset(id, mode, agg, ps.AggregateConfig, a.Allocator())
 	return t, d, nil
 }