Closes #3455 

~Requires https://github.com/influxdata/tdigest/pull/27 to merge first.~

- Uses `tdigest`'s new `ByteSizeForCompression` function to get the appropriate byte size to pass to the allocator.
- Moves creation of a digest from `QuantileAgg.Copy()` to `createQuantileTransformation` and simply resets the digest every time the aggregate is copied.
- Adds a NewQuantileAgg function for convenience in testing.

***Benchmarks before change:***
```
go test -bench=Quantile$ -run=^Benchmark -benchmem ./stdlib/universe
goos: linux
goarch: amd64
pkg: github.com/influxdata/flux/stdlib/universe
BenchmarkQuantile-24                   6         188348059 ns/op         1513738 B/op        137 allocs/op
PASS
ok      github.com/influxdata/flux/stdlib/universe      4.296s
```

***Benchmarks after change:***
```
go test -bench=Quantile$ -run=^Benchmark -benchmem ./stdlib/universe
goos: linux
goarch: amd64
pkg: github.com/influxdata/flux/stdlib/universe
BenchmarkQuantile-24                   6         184495440 ns/op           29280 B/op          8 allocs/op
PASS
ok      github.com/influxdata/flux/stdlib/universe      4.436s
```